refactor: split JwstDataDashboard into focused view components
Extract 7 components from the 1482-line god component into a
dashboard/ subdirectory with dedicated CSS files:
- DashboardToolbar (search, filters, view toggle, action buttons)
- DataCard (individual card in target/grouped view)
- LineageFileCard (file card within lineage level expansion)
- TargetGroupView (target-grouped display, uses DataCard)
- LineageView (lineage tree, uses LineageFileCard)
- DeleteConfirmationModal (unified obs + level delete modal)
- UploadModal (upload form)

Also extracts shared utilities:
- formatUtils.ts (formatFileSize)
- statusUtils.ts (getStatusColor)

Dashboard orchestrator reduced from 1482 to 603 lines (59%).
Pre-existing test infra failures (jsdom ESM compat) unchanged.

Closes #302
