+++ frontend/jwst-frontend/src/components/dashboard/DeleteConfirmationModal.css
@@ -1,0 +1,168 @@
+/* Delete Modal Overlay */
+.delete-modal-overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: var(--bg-overlay);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 1000;
+}
+
+/* Delete Modal */
+.delete-modal {
+  background-color: var(--bg-surface);
+  color: var(--text-primary);
+  padding: 24px;
+  border-radius: var(--radius-lg);
+  width: 90%;
+  max-width: 500px;
+  box-shadow: var(--shadow-lg);
+  max-height: 80vh;
+  overflow-y: auto;
+}
+
+.delete-modal h3 {
+  margin: 0 0 20px 0;
+  color: var(--color-error);
+  font-size: 20px;
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
+
+.delete-modal h3::before {
+  content: '\26A0\FE0F';
+}
+
+.delete-modal-content {
+  margin-bottom: 20px;
+}
+
+.delete-observation-id {
+  font-size: 14px;
+  color: var(--text-secondary);
+  margin-bottom: 8px;
+  word-break: break-all;
+}
+
+.delete-level-id {
+  font-size: 14px;
+  color: var(--text-secondary);
+  margin-bottom: 12px;
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
+
+.delete-level-id .level-badge {
+  display: inline-block;
+  padding: 2px 10px;
+  border-radius: 12px;
+  font-size: 12px;
+  font-weight: 600;
+  color: white;
+}
+
+.delete-summary {
+  font-size: 16px;
+  color: var(--text-primary);
+  margin-bottom: 16px;
+  padding: 10px;
+  background-color: var(--color-error-subtle);
+  border-radius: var(--radius-md);
+  border-left: 4px solid var(--color-error);
+}
+
+/* File List */
+.delete-file-list {
+  margin-bottom: 16px;
+  font-size: 13px;
+}
+
+.delete-file-list strong {
+  display: block;
+  margin-bottom: 8px;
+  color: var(--text-secondary);
+}
+
+.delete-file-list ul {
+  max-height: 150px;
+  overflow-y: auto;
+  margin: 0;
+  padding-left: 20px;
+  background-color: var(--bg-elevated);
+  border: 1px solid var(--border-default);
+  border-radius: var(--radius-md);
+  padding: 10px 10px 10px 30px;
+}
+
+.delete-file-list li {
+  color: var(--text-muted);
+  padding: 2px 0;
+  word-break: break-all;
+}
+
+/* Warning Message */
+.delete-warning {
+  color: #f87171;
+  font-size: 14px;
+  font-weight: 500;
+  padding: 12px;
+  background-color: var(--color-error-subtle);
+  border: 1px solid rgba(239, 68, 68, 0.3);
+  border-radius: var(--radius-md);
+  margin-top: 16px;
+}
+
+/* Modal Actions */
+.delete-modal-actions {
+  display: flex;
+  gap: 12px;
+  justify-content: flex-end;
+}
+
+.delete-cancel-btn {
+  padding: 10px 20px;
+  background-color: var(--bg-elevated);
+  color: var(--text-secondary);
+  border: 1px solid var(--border-default);
+  border-radius: var(--radius-md);
+  cursor: pointer;
+  font-size: 14px;
+  font-weight: 500;
+  transition: all var(--transition-fast);
+}
+
+.delete-cancel-btn:hover:not(:disabled) {
+  background-color: var(--bg-surface-hover);
+}
+
+.delete-cancel-btn:disabled {
+  opacity: 0.5;
+  cursor: not-allowed;
+}
+
+.delete-confirm-btn {
+  padding: 10px 20px;
+  background-color: var(--color-error);
+  color: white;
+  border: none;
+  border-radius: var(--radius-md);
+  cursor: pointer;
+  font-size: 14px;
+  font-weight: 500;
+  transition: all var(--transition-fast);
+}
+
+.delete-confirm-btn:hover:not(:disabled) {
+  background-color: #b91c1c;
+}
+
+.delete-confirm-btn:disabled {
+  opacity: 0.6;
+  cursor: not-allowed;
+}
