+++ frontend/jwst-frontend/src/components/dashboard/LineageFileCard.tsx
@@ -1,0 +1,104 @@
+import React from 'react';
+import { JwstDataModel } from '../../types/JwstDataTypes';
+import { getFitsFileInfo } from '../../utils/fitsUtils';
+import { getStatusColor } from '../../utils/statusUtils';
+import { API_BASE_URL } from '../../config/api';
+import './LineageFileCard.css';
+
+interface LineageFileCardProps {
+  item: JwstDataModel;
+  isSelectedForComposite: boolean;
+  onCompositeSelect: (dataId: string, event: React.MouseEvent) => void;
+  onView: (item: JwstDataModel) => void;
+  onProcess: (dataId: string, algorithm: string) => void;
+}
+
+const LineageFileCard: React.FC<LineageFileCardProps> = ({
+  item,
+  isSelectedForComposite,
+  onCompositeSelect,
+  onView,
+  onProcess,
+}) => {
+  const fitsInfo = getFitsFileInfo(item.fileName);
+  const canSelectForComposite = fitsInfo.viewable;
+
+  return (
+    <div className={`lineage-file-card ${isSelectedForComposite ? 'selected-composite' : ''}`}>
+      {fitsInfo.viewable && (
+        <div className="lineage-thumbnail">
+          {item.hasThumbnail ? (
+            <img
+              src={`${API_BASE_URL}/api/jwstdata/${item.id}/thumbnail`}
+              loading="lazy"
+              alt={item.fileName}
+              onError={(e) => {
+                (e.target as HTMLImageElement).style.display = 'none';
+              }}
+            />
+          ) : (
+            <span className="lineage-thumbnail-placeholder">🔭</span>
+          )}
+        </div>
+      )}
+      <div className="lineage-file-content">
+        <div className="file-header">
+          {fitsInfo.viewable && (
+            <button
+              className={`composite-select-btn small ${isSelectedForComposite ? 'selected' : ''}`}
+              onClick={(e) => onCompositeSelect(item.id, e)}
+              disabled={!canSelectForComposite}
+              title={
+                isSelectedForComposite ? 'Remove from composite selection' : 'Add to RGB composite'
+              }
+            >
+              {isSelectedForComposite ? '✓' : '+'}
+            </button>
+          )}
+          <span className="file-name" title={item.fileName}>
+            {item.fileName}
+          </span>
+          <div className="file-badges">
+            <span className={`fits-type-badge small ${fitsInfo.type}`} title={fitsInfo.description}>
+              {fitsInfo.viewable ? '🖼️' : '📊'}
+            </span>
+            {item.imageInfo?.filter && (
+              <span className="filter-badge small" title={`Filter: ${item.imageInfo.filter}`}>
+                {item.imageInfo.filter}
+              </span>
+            )}
+            <span
+              className={`status ${item.processingStatus}`}
+              style={{
+                color: getStatusColor(item.processingStatus),
+              }}
+            >
+              {item.processingStatus}
+            </span>
+          </div>
+        </div>
+        <div className="file-meta">
+          <span>Type: {item.dataType}</span>
+          <span>Size: {(item.fileSize / 1024 / 1024).toFixed(2)} MB</span>
+          {item.imageInfo?.observationDate && (
+            <span>Obs: {new Date(item.imageInfo.observationDate).toLocaleDateString()}</span>
+          )}
+          <span className="fits-type-label">{fitsInfo.label}</span>
+        </div>
+        <div className="file-actions">
+          <button
+            onClick={() => onView(item)}
+            className={!fitsInfo.viewable ? 'disabled' : ''}
+            disabled={!fitsInfo.viewable}
+            title={fitsInfo.viewable ? 'View FITS image' : fitsInfo.description}
+          >
+            {fitsInfo.viewable ? 'View' : 'Table'}
+          </button>
+          <button onClick={() => onProcess(item.id, 'basic_analysis')}>Analyze</button>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default LineageFileCard;
