+++ frontend/jwst-frontend/src/components/dashboard/TargetGroupView.css
@@ -1,0 +1,147 @@
+/* Grouped View */
+.grouped-view {
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+}
+
+.data-group {
+  margin-bottom: 30px;
+  background-color: var(--bg-surface);
+  border-radius: var(--radius-md);
+  padding: 15px;
+  border: 1px solid var(--border-default);
+}
+
+.group-header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  margin-bottom: 15px;
+  padding-bottom: 10px;
+  border-bottom: 1px solid var(--border-subtle);
+  cursor: pointer;
+  user-select: none;
+  transition: background-color var(--transition-fast);
+  border-radius: var(--radius-sm);
+  padding: 10px;
+  margin: -10px -10px 15px -10px;
+}
+
+.group-header:hover {
+  background-color: var(--bg-surface-hover);
+}
+
+.group-header:focus {
+  outline: 2px solid var(--accent-primary);
+  outline-offset: 2px;
+}
+
+.group-header-left {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+}
+
+.collapse-icon {
+  font-size: 12px;
+  color: var(--text-muted);
+  transition: transform var(--transition-fast);
+  width: 16px;
+  text-align: center;
+}
+
+.group-header h3 {
+  margin: 0;
+  color: var(--text-primary);
+  font-size: 18px;
+}
+
+/* Collapsed state */
+.data-group.collapsed {
+  padding-bottom: 5px;
+}
+
+.data-group.collapsed .group-header {
+  margin-bottom: 0;
+  border-bottom: none;
+}
+
+.group-count {
+  background-color: var(--bg-elevated);
+  padding: 4px 10px;
+  border-radius: 12px;
+  font-size: 12px;
+  color: var(--text-muted);
+  font-weight: 600;
+}
+
+.data-grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(350px, 1fr));
+  gap: 20px;
+}
+
+/* Grouped View Title and Meta */
+.group-title {
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+}
+
+.group-title h3 {
+  margin: 0;
+}
+
+.group-meta {
+  display: flex;
+  align-items: center;
+  flex-wrap: wrap;
+  row-gap: 4px;
+  gap: 8px;
+  font-size: 13px;
+  color: var(--text-muted);
+}
+
+.instrument-list {
+  color: #34d399;
+  font-weight: 500;
+}
+
+.filter-list {
+  color: #a78bfa;
+  font-weight: 500;
+}
+
+.meta-separator {
+  color: var(--text-faint);
+}
+
+.no-data {
+  grid-column: 1 / -1;
+  text-align: center;
+  padding: 60px 20px;
+  color: var(--text-secondary);
+}
+
+.no-data h3 {
+  margin-bottom: 10px;
+  color: var(--text-primary);
+}
+
+/* Focus States */
+.group-header:focus-visible {
+  outline: 2px solid var(--accent-primary);
+  outline-offset: 2px;
+}
+
+@media (max-width: 768px) {
+  .data-grid {
+    grid-template-columns: 1fr;
+  }
+
+  .group-meta {
+    flex-wrap: wrap;
+    row-gap: 4px;
+  }
+}
