+++ frontend/jwst-frontend/src/utils/formatUtils.ts
@@ -1,0 +1,6 @@
+export const formatFileSize = (bytes: number): string => {
+  if (bytes >= 1073741824) return `${(bytes / 1073741824).toFixed(2)} GB`;
+  if (bytes >= 1048576) return `${(bytes / 1048576).toFixed(2)} MB`;
+  if (bytes >= 1024) return `${(bytes / 1024).toFixed(2)} KB`;
+  return `${bytes} bytes`;
+};
